order logging

+++ lib/com.pjf.mat.core/src/main/java/com/pjf/mat/sys/DummyComms.java
@@ -182,7 +182,7 @@ public class DummyComms extends BaseComms implements Comms {
 	}
 
 	@Override
-	public void synchroniseClock(int syncOrigin) throws Exception {
+	public void synchroniseClock(long syncOrigin) throws Exception {
 		logger.info("synchroniseClock(" + syncOrigin + ")");		
 	}
